requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
django-ratelimit==4.1.0
requests==2.31.0
python-decouple==3.8
django-import-export==3.3.5